+++ Main/MemoryManager.hpp
@@ -34,12 +34,12 @@ namespace traceFileSimulator {
 
 class MemoryManager {
 public:
-	MemoryManager(int heapSize, int highWatermark, int collector, int traversal, int allocator);
+	MemoryManager(size_t heapSize, int highWatermark, int collector, int traversal, int allocator);
 	virtual ~MemoryManager();
 	//operations possible from trace file
-	int allocateObjectToRootset(int thread, int id, int size, int refCount, int classID);
-	int regionAllocateObjectToRootset(int thread, int id, int size, int refCount, int classID); //region-based; by Tristan
-	inline int postAllocateObjectToRootset(int thread, int id, int size, int refCount, int classID,void *address);  //added by Tristan
+	int allocateObjectToRootset(int thread, int id, size_t size, int refCount, int classID);
+	int regionAllocateObjectToRootset(int thread, int id, size_t size, int refCount, int classID); //region-based; by Tristan
+	inline int postAllocateObjectToRootset(int thread, int id, size_t size, int refCount, int classID,void *address);  //added by Tristan
 	int requestRootDelete(int thread, int id);
 	int requestRootAdd(int thread, int id);
 	inline int preSetPointer(int thread, int parentID, int parentSlot, int childID); //added by Tristan
@@ -74,15 +74,15 @@ public:
 
 private:
 	bool isAlreadyRoot(int thread, int id);
-	int* computeHeapsizes(int heapSize);
-	void initAllocators(int heapsize);
+	size_t* computeHeapsizes(size_t heapSize);
+	void initAllocators(size_t heapsize);
 	void initContainers();
 	void initGarbageCollectors(int highWatermark);
-	void *allocate(int size, int generation);
-	void *allocate(int size, int generation, int thread);
+	void *allocate(size_t size, int generation);
+	void *allocate(size_t size, int generation, int thread);
 	void addRootToContainers(Object* object, int thread);
 	void addToContainers(Object* object);
-	void *shift(int size);
+	void *shift(size_t size);
 	
 	allocatorEnum _allocator;
 	collectorEnum _collector;
